Fix attempt counting and logic
Only start a new attempt when at least one move is made.
Only count attempts if at least one move was made.

Fix when the user goes to menu while playing an attempt of a daily
deal and returns to the same deal through the daily deal menu it
would reset the deal.